Restructure: class name changes.
