+++ src/main/java/com/jfrog/ide/common/tree/IssueNode.java
@@ -1,168 +1,18 @@
 package com.jfrog.ide.common.tree;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import org.apache.commons.lang3.StringUtils;
+import javax.swing.tree.TreeNode;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
+public abstract class IssueNode extends ComparableSeverityTreeNode implements SubtitledTreeNode {
 
-import static com.jfrog.ide.common.tree.Severity.getNotApplicableSeverity;
-
-/**
- * @author yahavi
- */
-public class IssueNode extends VulnerabilityOrViolationNode {
-
-    private Severity severity = Severity.Normal;
-    private List<String> fixedVersions;
-    private List<String> infectedVersions;
-    private List<String> references;
-    private Cve cve;
-    private String summary;
-    private String issueId;
-    private String lastUpdated;
-    private List<String> watchNames;
-    private ResearchInfo researchInfo;
-    private List<ApplicableIssueNode> applicableIssues;
-
-    public IssueNode() {
+    public String getIcon() {
+        return getSeverity().getIconName();
     }
 
-    @SuppressWarnings("unused")
-    public IssueNode(String issueId, Severity severity, String summary, List<String> fixedVersions, List<String> infectedVersions,
-                     Cve cve, String lastUpdated, List<String> watchNames, List<String> references, ResearchInfo researchInfo) {
-        this.issueId = issueId;
-        this.severity = severity;
-        this.summary = summary;
-        this.fixedVersions = fixedVersions;
-        this.infectedVersions = infectedVersions;
-        this.cve = cve;
-        this.lastUpdated = lastUpdated;
-        this.watchNames = watchNames;
-        this.references = references;
-        this.researchInfo = researchInfo;
-    }
-
-    public String getIssueId() {
-        return this.issueId;
-    }
-
-    public Severity getSeverity() {
-        return isApplicable() != null && !isApplicable() ? getNotApplicableSeverity(severity): severity;
-    }
-
-    public Severity getSeverity(boolean masked) {
-        if (!masked) {
-            return this.severity;
+    public DependencyNode getParentArtifact() {
+        TreeNode parent = getParent();
+        if (!(parent instanceof DependencyNode)) {
+            return null;
         }
-        return getSeverity();
+        return (DependencyNode) parent;
     }
-
-    @SuppressWarnings("unused")
-    public String getSummary() {
-        return summary;
-    }
-
-    @SuppressWarnings("unused")
-    public List<String> getFixedVersions() {
-        return fixedVersions;
-    }
-
-    @SuppressWarnings("unused")
-    public void setFixedVersions(List<String> fixedVersions) {
-        this.fixedVersions = fixedVersions;
-    }
-
-    @SuppressWarnings("unused")
-    public List<String> getInfectedVersions() {
-        return infectedVersions;
-    }
-
-    @SuppressWarnings("unused")
-    public List<String> getReferences() {
-        return references;
-    }
-
-    public Cve getCve() {
-        return cve;
-    }
-
-    public String getLastUpdated() {
-        return lastUpdated;
-    }
-
-    public List<String> getWatchNames() {
-        return watchNames;
-    }
-
-    @SuppressWarnings("unused")
-    public ResearchInfo getResearchInfo() {
-        return researchInfo;
-    }
-
-    @JsonIgnore
-    @SuppressWarnings("WeakerAccess")
-    public boolean isTopSeverity() {
-        return getSeverity() == Severity.Critical;
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (!(other instanceof IssueNode)) {
-            return false;
-        }
-        IssueNode otherIssue = (IssueNode) other;
-        return StringUtils.equals(otherIssue.getIssueId(), getIssueId()) &&
-                StringUtils.equals(otherIssue.getCveIdOrEmpty(), getCveIdOrEmpty());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(issueId, cve);
-    }
-
-    @Override
-    public String getTitle() {
-        return StringUtils.firstNonBlank(getCveIdOrEmpty(), issueId);
-    }
-
-    @Override
-    public String getSubtitle() {
-        return null;
-    }
-
-    @Override
-    public String toString() {
-        return getTitle();
-    }
-
-    private String getCveIdOrEmpty() {
-        if (cve == null) {
-            return "";
-        }
-        return cve.getCveId();
-    }
-
-    public List<ApplicableIssueNode> getApplicableIssues() {
-        return applicableIssues;
-    }
-
-    public void setApplicableIssues(List<ApplicableIssueNode> applicableIssues) {
-        this.applicableIssues = applicableIssues;
-    }
-
-    public void AddApplicableIssues(ApplicableIssueNode issue) {
-        this.applicableIssues = this.applicableIssues == null ? new ArrayList<>() : this.applicableIssues;
-        this.applicableIssues.add(issue);
-    }
-
-    /**
-     * Returns true if the issue is applicable, false if not.
-     * If the applicability status is unknown, null will be returned.
-     */
-    public Boolean isApplicable() {
-        return this.applicableIssues != null ? this.applicableIssues.size() > 0 : null;
-    }
-
 }
