--- src/main/java/com/jfrog/ide/common/tree/VulnerabilityOrViolationNode.java
@@ -1,18 +1,0 @@
-package com.jfrog.ide.common.tree;
-
-import javax.swing.tree.TreeNode;
-
-public abstract class VulnerabilityOrViolationNode extends ComparableSeverityTreeNode implements SubtitledTreeNode {
-
-    public String getIcon() {
-        return getSeverity().getIconName();
-    }
-
-    public DependencyNode getParentArtifact() {
-        TreeNode parent = getParent();
-        if (!(parent instanceof DependencyNode)) {
-            return null;
-        }
-        return (DependencyNode) parent;
-    }
-}
